requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
email-validator==2.1.1
python-dotenv==1.0.1
sqlalchemy==2.0.27
pymysql==1.1.0 